refactor: move abstract cms functionality into cms class
