+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/const/AbstractConstMultiSetImpl.kt
@@ -14,41 +14,23 @@ internal interface AbstractConstMultiSetImpl<E> {
     val counts: CountsMap<E>
     val size: Int
 
-    fun plus(other: MultiSet<E>): MultiSet<E> {
-        return combineCounts(counts, other, Int::plus, true, const = false)
-    }
-    fun minus(other: MultiSet<E>): MultiSet<E> {
-        return combineCounts(counts, other, Int::minus, false, const = false)
-    }
-    fun intersect(other: MultiSet<E>): MultiSet<E> {
-        return combineCounts(counts, other, ::min, false, const = false)
-    }
-
-    fun plusC(other: ConstMultiSet<E>): ConstMultiSet<E> {
-        return combineCounts(counts, other, Int::plus, true, const = true) as ConstMultiSet<E>
-    }
-    fun minusC(other: ConstMultiSet<E>): ConstMultiSet<E> {
-        return combineCounts(counts, other, Int::minus, false, const = true) as ConstMultiSet<E>
-    }
-    fun intersectC(other: ConstMultiSet<E>): ConstMultiSet<E> {
-        return combineCounts(counts, other, ::min, false, const = true) as ConstMultiSet<E>
-    }
-
-    // check equality to another multiset
-    fun equalsSet(other: MultiSet<*>): Boolean {
-        return when (other) {
-            is AbstractMultiSetImpl<*> -> counts == CountsMap.from(other)
-            is AbstractConstMultiSetImpl<*> -> counts == other.counts
-            else -> {
-                @Suppress(Suppressions.UNCHECKED_CAST)
-                other as MultiSet<E>
-                size == other.size && counts.distinct.all { counts.getCountOf(it) == other.getCountOf(it) }
-            }
-        }
-    }
-
-    fun getHashCode(): Int {
-        val hashCode = counts.hashCode()
-        return 31 * hashCode + MultiSet::class.java.name.hashCode()
-    }
+//    fun plus(other: MultiSet<E>): MultiSet<E> {
+//        return combineCounts(counts, other, Int::plus, true, const = false)
+//    }
+//    fun minus(other: MultiSet<E>): MultiSet<E> {
+//        return combineCounts(counts, other, Int::minus, false, const = false)
+//    }
+//    fun intersect(other: MultiSet<E>): MultiSet<E> {
+//        return combineCounts(counts, other, ::min, false, const = false)
+//    }
+//
+//    fun plusC(other: ConstMultiSet<E>): ConstMultiSet<E> {
+//        return combineCounts(counts, other, Int::plus, true, const = true) as ConstMultiSet<E>
+//    }
+//    fun minusC(other: ConstMultiSet<E>): ConstMultiSet<E> {
+//        return combineCounts(counts, other, Int::minus, false, const = true) as ConstMultiSet<E>
+//    }
+//    fun intersectC(other: ConstMultiSet<E>): ConstMultiSet<E> {
+//        return combineCounts(counts, other, ::min, false, const = true) as ConstMultiSet<E>
+//    }
 }
