+++ kotlin-utils/src/main/kotlin/xyz/lbres/kotlinutils/set/multiset/const/ConstMutableMultiSetImpl.kt
@@ -1,11 +1,10 @@
 package xyz.lbres.kotlinutils.set.multiset.const
 
-import xyz.lbres.kotlinutils.set.multiset.MultiSet
 import xyz.lbres.kotlinutils.set.multiset.utils.CountsMap
 import kotlin.math.min
 
 // final implementation of ConstMutableMultiSet
-internal class ConstMutableMultiSetImpl<E>(initialElements: Collection<E>) : ConstMutableMultiSet<E>(), AbstractConstMultiSetImpl<E> {
+internal class ConstMutableMultiSetImpl<E>(initialElements: Collection<E>) : ConstMutableMultiSet<E>() {
     /**
      * If all properties are up-to-date with the most recent changes to the counts map
      */
@@ -96,22 +95,6 @@ internal class ConstMutableMultiSetImpl<E>(initialElements: Collection<E>) : Con
         _size = 0
         allPropertiesUpdated = false
     }
-
-    override fun getCountOf(element: E): Int = counts.getCountOf(element)
-    override fun contains(element: E): Boolean = counts.contains(element)
-    override fun containsAll(elements: Collection<E>): Boolean = counts.containsAll(elements)
-
-    override fun plus(other: MultiSet<E>): MultiSet<E> = super<AbstractConstMultiSetImpl>.plus(other)
-    override fun minus(other: MultiSet<E>): MultiSet<E> = super<AbstractConstMultiSetImpl>.minus(other)
-    override fun intersect(other: MultiSet<E>): MultiSet<E> = super<AbstractConstMultiSetImpl>.intersect(other)
-
-    override fun plusC(other: ConstMultiSet<E>): ConstMultiSet<E> = super.plusC(other)
-    override fun minusC(other: ConstMultiSet<E>): ConstMultiSet<E> = super.minusC(other)
-    override fun intersectC(other: ConstMultiSet<E>): ConstMultiSet<E> = super.intersectC(other)
-
-    override fun isEmpty(): Boolean = counts.isEmpty()
-    override fun equals(other: Any?): Boolean = other is MultiSet<*> && equalsSet(other)
-    override fun hashCode(): Int = getHashCode()
 
     override fun iterator(): MutableIterator<E> {
         updateMutableValues()
